pyproject.toml
[project]
name = "spotart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rgbmatrixemulator>=0.16.3",
]
